requirements.txt
# requirements.txt
nba-api==1.4.1
opencv-python==4.10.0.84
boto3==1.34.144
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
